remove the ocaml version constraint from the opam file
This version of the package includes a compatibility version of the
Seq module, and so compiles on 4.07+ as well.

When installing with opam, it will not be selected as seq.base will
be used.

However, when installing as a duniverse monorepo, it is useful to
have this repository directly used, since opam is not present to
install the `seq` dummy file.

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>
Co-authored-by: @NathanReb